修复类型不匹配警告

+++ Classes/FanCalculator/FanCalculatorScene.cpp
@@ -100,7 +100,7 @@ cocos2d::Node *createFanResultNode(const mahjong::fan_table_t &fan_table, int fo
         widget->setContentSize(label->getContentSize());
         node->addChild(widget);
         widget->addClickEventListener([j](Ref *) {
-            Director::getInstance()->pushScene(FanDefinitionScene::create(j));
+            Director::getInstance()->pushScene(FanDefinitionScene::create(static_cast<size_t>(j)));
         });
     }
